UWP: SDL bridge build mode
The Xbox UWP environment doesn't really support windows. The SDL bridge
serves as a faux replacement allowing for windows of any size to be
created and trigger a resize of the final buffer.  This allows for direct
rendering of lower resolutions without needing to add a framebuffer to
existing applications.

tldr: Fixes the "poststamp" issue SDL-UWP has on xbox when targeting
resolutions lower than native.